+++ src/Client.php
@@ -105,15 +105,8 @@ abstract class Client
 
     /**
      * Get a class instance throwing an exception if check fails
-     *
-     * @param string|null     $param1   path or host
-     * @param string|int|null $param2   Java binary path or port for web client
-     * @param array           $options  Java args or options for cURL request
-     * @param bool            $check    check JAR file or server connection
-     * @return \Vaites\ApacheTika\Clients\CLI|\Vaites\ApacheTika\Clients\REST
-     * @throws \Exception
      */
-    public static function make(string $param1 = null, $param2 = null, array $options = null, bool $check = null): Client
+    public static function make(string $param1 = null, $param2 = null, array $options = null, bool $check = null): CLI|REST
     {
         if($param1 !== null && preg_match('/\.jar$/', $param1))
         {
@@ -129,14 +122,8 @@ abstract class Client
 
     /**
      * Get a class instance delaying the check
-     *
-     * @param string|null $param1 path or host
-     * @param int|null    $param2 Java binary path or port for web client
-     * @param array       $options options for cURL request
-     * @return \Vaites\ApacheTika\Clients\CLI|\Vaites\ApacheTika\Clients\REST
-     * @throws \Exception
      */
-    public static function prepare($param1 = null, $param2 = null, $options = []): Client
+    public static function prepare($param1 = null, $param2 = null, $options = []): CLI|REST
     {
         return self::make($param1, $param2, $options, false);
     }
@@ -434,7 +421,6 @@ abstract class Client
             throw new Exception("Apache Tika $version is unsupported");
         }
 
-        $this->checked = true;
         $this->version = $version;
 
         return $this;
@@ -715,12 +701,12 @@ abstract class Client
     /**
      * Check Java binary, JAR path or server connection
      */
-    abstract public function check(): void;
+    abstract protected function check(): void;
 
     /**
      * Configure and make a request and return its results.
      *
      * @throws \Exception
      */
-    abstract public function request(string $type, string $file = null): ?string;
+    abstract protected function request(string $type, string $file = null): ?string;
 }